+++ pylabrobot/heating_shaking/__init__.py
@@ -4,4 +4,4 @@ from pylabrobot.heating_shaking.backend import HeaterShakerBackend
 from pylabrobot.heating_shaking.chatterbox import HeaterShakerChatterboxBackend
 from pylabrobot.heating_shaking.hamilton_backend import HamiltonHeaterShakerBackend
 from pylabrobot.heating_shaking.heater_shaker import HeaterShaker
-from pylabrobot.heating_shaking.inheco import InhecoThermoShakeBackend
+from pylabrobot.heating_shaking.inheco_backend import InhecoThermoShakeBackend
